Player UI, Runes, MeleeAttack and Specials Melee attack.
Muchos cambios hechos en este commit, entre los mas importantes se encuentra el cambio de la UI del player, tambien incluir runas que seran la dinamica de combate escogia como energia.

+++ 2D-Rpg-Game/Assets/Scripts/EnemyScript/EnemyAttack.cs
@@ -12,7 +12,7 @@ public class EnemyAttack : MonoBehaviour
     int _animationHash;
     float _timer;
     float _coolDown;
-    float _damage;
+    int _damage;
     bool _playerInRange;
 
     Animator anim;
